Liquor-backend/src/core_service/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
confluent-kafka==2.3.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
